+++ generator/integration-tests/test_env.dart
@@ -1,4 +1,6 @@
 import 'dart:io';
+import 'package:objectbox/objectbox.dart';
+export 'package:objectbox/objectbox.dart';
 import 'package:objectbox/internal.dart';
 export 'package:objectbox/internal.dart';